CP-56 complete

+++ src/main/java/ee/bcs/carportal/service/car/CarService.java
@@ -1,5 +1,7 @@
 package ee.bcs.carportal.service.car;
 
+import ee.bcs.carportal.infrastructure.exception.DatabaseConflictException;
+import ee.bcs.carportal.infrastructure.exception.ResourceNotFoundException;
 import ee.bcs.carportal.persistence.car.Car;
 import ee.bcs.carportal.persistence.car.CarMapper;
 import ee.bcs.carportal.persistence.car.CarRepository;
@@ -25,7 +27,7 @@ public class CarService {
     private final CarMapper carMapper;
 
     public CarInfo findCarInfo(Integer carId) {
-        Car car = carRepository.getReferenceById(carId);
+        Car car = carRepository.findById(carId).orElseThrow(() -> new ResourceNotFoundException("Resource not found"));
         return carMapper.toCarInfo(car);
     }
 
@@ -43,6 +45,10 @@ public class CarService {
     }
 
     public void addCar(CarDto carDto) {
+        boolean carExists = carRepository.carExistsBy(carDto.getManufacturerId(), carDto.getModel(), carDto.getYear());
+        if (carExists) {
+            throw new DatabaseConflictException("Car already exists");
+        }
         Manufacturer manufacturer = manufacturerRepository.findById(carDto.getManufacturerId()).orElse(null);
         FuelType fuelType = fuelTypeRepository.findById(carDto.getFuelTypeId()).orElse(null);
 
@@ -53,27 +59,28 @@ public class CarService {
     }
 
     public CarDetailedInfo findCarDetailedInfo(Integer carId) {
-        Car car = carRepository.findById(carId)
-                .orElseThrow(() -> new RuntimeException("Car not found with id: " + carId));
-
+        Car car = carRepository.findById(carId).orElseThrow(() -> new ResourceNotFoundException("Resource not found"));
         return carMapper.toCarDetailedInfo(car);
     }
 
     public void updateCar(Integer carId, CarDto carDto) {
-        Car car = carRepository.getReferenceById(carId);
+        Car car = carRepository.findById(carId).orElseThrow(() -> new ResourceNotFoundException("Resource not found"));
         carMapper.updateCar(carDto, car);
         if (!car.getManufacturer().getId().equals((carDto.getManufacturerId()))) {
-            Manufacturer manufacturer = manufacturerRepository.getReferenceById(carDto.getManufacturerId());
+            Manufacturer manufacturer = manufacturerRepository.findById(carDto.getManufacturerId()).orElseThrow(() -> new ResourceNotFoundException("Resource not found"));
             car.setManufacturer(manufacturer);
         }
         if (!car.getFuelType().getId().equals(carDto.getFuelTypeId())) {
-            FuelType fuelType = fuelTypeRepository.getReferenceById(carDto.getFuelTypeId());
+            FuelType fuelType = fuelTypeRepository.findById(carDto.getFuelTypeId()).orElseThrow(() -> new ResourceNotFoundException("Resource not found"));
             car.setFuelType(fuelType);
         }
         carRepository.save(car);
     }
 
     public void deleteCar(Integer carId) {
+        if (!carRepository.existsById(carId)) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
         carRepository.deleteById(carId);
     }
 }
